refactor setMessage method. Clean up and move view definition outside
The message view is now self-contianed and can function outside of the scope of the setMessage method. Message defaults are a constant, and use them when initializing a new message view. Clone the defaults when they are used in the setMessage method so the values can be adjusted based on the method options.