when user adjusts offset, the whole duration of the item should "slide"
That is, as the user offsets the start time from the parent, if the duration is less than that of the parent both the start and end times should be adjusted so that the duration is maintained. This will mean that the whole child item is offset, and the duration is maintained until it brings up against the limit of the parent's end time.